+++ release-notes.md
@@ -6,6 +6,7 @@
 - Add support for wp-drm-lease-v1.
 - Focus-follows-mouse can now be disabled.
 - Add support for pointer-gestures-unstable-v1.
+- Configs can now handle switch events (laptop lid closed/opened).
 
 # 1.1.0 (2024-04-22)